requirements.txt
streamlit
streamlit-cropper
rembg
diffusers
torch
Pillow
numpy
opencv-python
zipfile36
tqdm